działa dla pierwszych 50 posłów       +obsługa wyjazdów       +włączenie wyjazdów w koszta       TODO       +współbieżnie (jeśli się uda)       +testy

+++ src/com/company/Cost.java
@@ -4,6 +4,6 @@ package com.company;
  * Created by Gabrysia on 16.12.2016.
  */
 abstract public class Cost {
-    protected float cost;
-    public float getCost(){ return cost;}
+    protected double cost;
+    public double getCost(){ return cost;}
 }
